# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Leerr API
  version: "0.2"
  description: |
    Same-origin JSON API for the Leerr web app.

    Authentication: web sessions use an HttpOnly, SameSite=Strict cookie
    (`__Host-leerr` over HTTPS) and must send `X-CSRF-Token` on every
    non-GET request. Native sessions send `Authorization: Bearer <token>`
    and need no CSRF token. Each session kind only works through its own channel.

    Every error has the body `{"error": {"code": "...", "message": "..."}}`;
    messages are written for end users.
servers:
  - url: /
security:
  - cookie: []
  - bearer: []
...
paths:
...
  /api/v1/setup:
    get:
      tags: [Sessions]
      summary: Whether first-run setup is still required
      security: []
      responses:
        "200":
          description: Setup state
          content:
            application/json:
              schema:
                type: object
                required: [required, fixturePreview]
                properties:
                  required: { type: boolean }
                  fixturePreview: { type: boolean, description: True for the in-memory demo server. }
    post:
      tags: [Sessions]
      summary: Create the first administrator using the one-time setup token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, username, password]
              properties:
                token: { type: string }
                username: { $ref: "#/components/schemas/Username" }
                password: { $ref: "#/components/schemas/Password" }
      responses:
        "200": { $ref: "#/components/responses/Empty" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/sessions:
    post:
      tags: [Sessions]
      summary: Sign in
      description: Rate limited per address, and locked per account after repeated failures.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
                device: { type: string, enum: [web, native], default: web }
                name: { type: string, maxLength: 80, default: Browser }
      responses:
        "200":
          description: Signed in. Web sessions also set the session cookie.
          content:
            application/json:
              schema:
                type: object
                required: [user, csrf]
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  csrf: { type: [string, "null"] }
                  token: { type: string, description: Native sessions only. }
        "401": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
    get:
      tags: [Sessions]
      summary: The current user's signed-in devices
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [id, name, device, createdAt, expiresAt, current]
                      properties:
                        id: { type: string }
                        name: { type: string }
                        device: { type: string, enum: [web, native] }
                        createdAt: { type: string, format: date-time }
                        expiresAt: { type: string, format: date-time }
                        current: { type: boolean }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/sessions/{id}:
    delete:
      tags: [Sessions]
      summary: Sign out a session (`current` for this one); its streams stop immediately
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { $ref: "#/components/responses/Empty" }
        "404": { $ref: "#/components/responses/Error" }
...
  /api/v1/streams/{ticket}:
    get:
      tags: [Library]
      summary: The original audio file, with Range support
      security: []
      parameters:
        - { name: ticket, in: path, required: true, schema: { type: string } }
        - { name: Range, in: header, schema: { type: string } }
        - { name: If-Range, in: header, schema: { type: string } }
      responses:
        "200": { description: Whole file, content: { audio/*: { schema: { type: string, format: binary } } } }
        "206": { description: Requested range, content: { audio/*: { schema: { type: string, format: binary } } } }
        "401": { $ref: "#/components/responses/Error" }
        "416": { description: Range not satisfiable }
        "429": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    cookie: { type: apiKey, in: cookie, name: __Host-leerr }
    bearer: { type: http, scheme: bearer }
  responses:
    Empty:
      description: Success
      content:
        application/json:
          schema: { type: object }
    Error:
      description: Error
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: object
                required: [code, message]
                properties:
                  code: { type: string }
                  message: { type: string }
  schemas:
    Username: { type: string, minLength: 1, maxLength: 64, pattern: "^[a-zA-Z0-9_.-]+$" }
    Password: { type: string, minLength: 10, maxLength: 256 }
    Role: { type: string, enum: [admin, member] }
    User:
      type: object
      required: [id, username, role]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        role: { $ref: "#/components/schemas/Role" }
...